WIP: Device layer port to nRF52840
-- Introduced new GenericConnectivityManagerImpl_Thread<> class.  The new class provides
generic implementations of ConnectivityManager features that are specific to devices that
support Thread. The class is designed to be independent of the particular Thread stack in
use (implying, for example, that the code does not make direct use of any OpenThread APIs).
This is achieved by delegating all stack-specific operations to the ThreadStackManager
class.

The GenericConnectivityManagerImpl_Thread<> class is currently a work in progress and will
grow in functionality as Thread support in the Device Layer is fleshed out.

-- Introduced new GenericConnectivityManagerImpl_NoThread<> class which implements
Thread-specific ConnectivityManager features for platforms that don’t have a Thread radio.

-- Implemented logic to detect service connectivity via Thread by detecting the presence of
Thread border routers offering routes for the Weave service subnet.  Typically, these routes
will consist of /48 routes for the Weave fabric prefix that are being advertised by Weave
border routers.

When the Device Layer detects the introduction or loss of a service route, it raises a
ServiceConnectivityChange event, which can then be used by the application and/or other
Device Layer components to trigger specific behaviors (e.g. a service time sync operation by
the TimeSyncManager).

-- Added new service connectivity methods to ConnectivityManager:
HaveServiceConnectivityViaTunnel() and HaveServiceConnectivityViaThread().

-- Added ThreadStackManager::HaveRouteToAddress().  This method takes a destination IPv6
address and determines whether a route exists via the Thread interface to the specified
address.

-- Re-organized generic ThreadStackManager implementation classes:

The GenericThreadStackManagerImpl_OpenThread<> class provides a generic implementation of
ThreadStackManager features that works in conjunction with OpenThread, but is independent of
host OS and IP stack.

The GenericThreadStackManagerImpl_OpenThread_LwIP<> class inherits from
GenericThreadStackManagerImpl_OpenThread<> and provides features that are specific to
OpenThread when used with LwIP as the host IP stack.

-- Moved template instantiation directives into generic implementation headers.

-- Made general functions for logging OpenThread packets and state changes.