nRF52840 adaptation: J-Link monitor mode debugging
-- Enabled the use of J-Link monitor mode debugging in the bring-up app.

The nRF52840 bring-up app can now be built with support for the J-Link monitor mode debugging feature.  Monitor
mode debugging works in conjunction with the J-Link GDB server to allow parts of the system to continue running
while the target device is ‘stopped’ at a breakpoint.  Because of the way the Nordic SoftDevice works, this is
the only viable way to debug an application that incorporates the SoftDevice.  Unfortunately, J-Link monitor
mode requires use of the *ludicrously* expensive J-Link Plus debug probe, making the on-board debugger
interface useless for this purpose.

-- Updated the start-gdb.sh script to automatically enable monitor mode debugging whenever it detects that the
application has been built with the appropriate support.

-- Updated README for the bring-up app to explain the use of monitor mode debugging.