WIP: Device layer port to nRF52840
-- Working implementation of core Network Provisioning features for Thread.

As of this commit, the Device Layer supports the following Network Provisioning commands for
Thread networks: AddNetwork, RemoveNetwork, UpdateNetwork, EnableNetwork, DisableNetwork,
GetNetworks and TestNetwork.  Support for ScanNetworks and SetRendezvousMode remains a work
in progress.

-- Extended ConnectivityManager with APIs for controlling Thread connectivity. Enhanced
GenericConnectivityManagaerImpl_Thread<> and GenericConnectivityManagaerImpl_NoThread<> to
support these new APIs.

-- Extended ThreadStackManager with APIs that provide implementation-neutral access to
Thread stack control functions.  These allow other Device Layer components such as the
generic network provisioning server implementation to be independent from the choice of
Thread stack.

-- Enhanced GenericThreadStackManagerImpl_OpenThread<> to support new ThreadStackManager
APIs.

-- Restructured GenericNetworkProvisioningServerImpl<> to support both WiFi and Thread,
enabled separately at compile time.

-- Revised NetworkInfo class to support Thread configuration data.  As part of this, support
was added for specifying a Thread PSKc.  This code is preliminary pending definition of the
TLV encoding for the PSKc parameter.

-- Implemented TestNetwork for Thread networks, using the presence of another router in the
mesh as evidence of Thread network connectivity.

-- Created a generalized infrastructure for TestNetwork that makes it easier to send back
detailed status reports.

-- Added ChildNodesChanged flags to generic ThreadConnectivityChange event to signal when
there is a change in the set of children attached to the node.

-- Implemented support for generating random values for any Thread network parameters not
supplied by the client in an AddNetwork or UpdateNetwork command.