nRF52840 adaptation: Continued work on debugging support
-- Modified the start-jlink-gdb-server.sh script to automatically load the SEGGER FreeRTOS
GDB extension.  This makes it possible to see and interact with FreeRTOS tasks in GDB.

-- Added a FreeRTOSDebugConfig structure to the bring-up app.  FreeRTOSDebugConfig is a
global data structure containing FreeRTOS configuration information that is used by the
SEGGER FreeRTOS GDB extension.  The structure is defined in a .h file with a pre-defined
name that is included directly in FreeRTOS's tasks.c source file by means of a configuration
option.

-- Cleaned up the naming of threads in the bring-up app.

-- Disabled FreeRTOS timers, and the FreeRTOS timer task, as these aren’t used in the
bring-up app.