Generalized the Weave error-to-string facility
-- Re-implemented Weave’s ErrorStr() function to use a modular design where the work of
translating an error to string form is delegated to a series of formatting functions which
can be registered at runtime.  The new design makes it possible for error formatting
functionality to be implemented separately for each category of error.  Additionally,
adapters of Weave code can install their own error-to-string translations for errors that
are unique to a particular platform or execution environment.

Once landed, this feature will be used to provide SDK-specific error translations for the
nRF52840 and ESP32 platforms.

-- Using the above facility, error translations for the various OpenWeave-defined error
categories were moved out of ErrorStr.cpp and into their own files.

-- Minor adjustments were made to the output format of error strings to make them more
consistent and more informative.  Notably, long-format error strings now include the error
code in both decimal and hex format.

-- Reduced memory usage when using long-format error strings by decreasing the size of the
static buffer used to error strings from 1000 to 256 bytes.  The new size seems to be more
than adequate for the existing crop of error translations.

-- Added a config option to override the print format used for error values in short-form
error strings.  This makes it possible, for example, to have error values printed in decimal
instead of hex.