WIP: Device layer port to nRF52840
-- Restructured the Thread related Device Layer events.  ThreadStateChange indicates a
change in the state of the Thread stack (role changes, address changes, child changes, etc.)
while ThreadConnectivityChange indicates a change in the link state of the host-side (i.e.
LwIP) Thread interface.  ThreadStateChange is now independent of the particular Thread
stack, but also conveys the OpenThread state change flags on systems that use OpenThread.

-- Arranged for MessageLayer RefreshEndpoints() to be called immediately whenever a change
is made to the state or addresses associated with the LwIP Thread interface.  This ensures
the MessageLayer listening state is updated before other components react to the change.

-- Modified the generic network provisioning server code to derive the Thread mesh prefix
from the extended PAN id whenever the mesh prefix is not specified on an AddNetwork.  This
pattern follows a similar convention within OpenThread itself.

-- Log additional information in LogOpenThreadStateChange().