nRF52840 adaptation: New makefiles for bring-up app.
-- Rewrote the Makefile system for the bring-up app.  The new system is cleaner and easier
to understand, making it better suited as a template for creating example or production
applications.  The new Makefile has the same command-line interface, and so should be
compatible with existing build processes.

-- Introduced nrf5-app.mk, which is a set of default make constructs for building
applications using Nordic’s SDK.  This file is an easier-to-use alternative to Nordic’s
Makefile.common example.  nrf5-app.mk offers general build rule construction, dependency
generation, support for standard build options, support for device flashing and built-in
help.  It is also well commented.

-- Rewrote openweave.mk (now called nrf5-openweave.mk) to work with nrf5-app.mk.  With this
change, incorporating OpenWeave in an application is as simple as including
nrf5-openweave.mk at the top of the Makefile.  As before, nrf5-openweave.mk takes care of
calling out to OpenWeave’s build system.