Additional crypto flexibility in OpenWeave
-- Added support mbed TLS as an underlying implementation for AES128/AES256 and SHA1/SHA256
in OpenWeave.  This configuration is convenient in situations where OpenWeave is built
together with OpenThread.

Note that the ability to use mbed TLS’s elliptic curve implementation is not available at
this time.

-- Generalize support for build-time selection of external AES and SHA implementations.
It’s now more straightforward to enable custom implementations and control the inclusion of
external headers during the build process.