Fixes to the new logging configuration
The new logging formatting code decentralized error formatting:
different Weave sub-libraries -- WeaveSystem, Inet, Warm -- now
provide their own error formatting.  Within Weave, however, it is
desirable to have a single config flag that switches between short and
verbose error codes.  To that end, we retain the control over the
logging format via WEAVE_CONFIG_SHORT_ERR_STRING, and to ensure the
propagation of that define into appropriate modules, we include the
config headers that could impact the value of that define in the
SystemConfig.h.

Include <stddef.h> in the new error formatter files, to ensure that
NULL is defined.