Device Layer: Support for User Selected Mode
-- Implemented support in the Device Layer for User Selected Mode.

User Selected Mode is a feature of the Weave Device Identify protocol. When a device is
placed in User Selected Mode, it will respond to Device Identify requests that have the
UserSelectedMode flag set.  A device not in User Selected Mode will ignore such requests.
User Selected Mode is typically initiated by a button press or some other form of direct
user interaction.  The User Selected Mode feature allows a client, be it a mobile
application or another device, to quickly identify a device that has been singled out by the
user.  A device will automatically deactivate User Selected Mode after a configurable period
of time.  Alternatively, the mode can be expressly deactivated by the application.

Since device UI is the responsibility of the application, the Device Layer exposes APIs (via
the ConnectivityMgr singleton) for activating and deactivating User Selected Mode, and for
configuring the mode timeout period.  These must be called by the application at the
appropriate times.

-- Introduced the GenericConnectivityManagerImpl<> template class, which provides default
implementations for the User Selected Mode APIs on the ConnectivityManager interface.  These
methods merely delegate the calls to the DeviceDescriptionServer server, which handles all
the logic.