Clean up build support for OS X
-- Revised Makefile-Standalone to automatically detect and use the standard OpenSSL package
installed on either Linux or OS X, when available.  On Linux, the logic uses pkg-config to
determine if the standard OpenSSL development package (typically libssl-dev) has been
installed.  On OS X, the logic uses homebrew to look for the OpenSSL formula.  If a standard
package is not detected, the build falls-back to using the internal version of OpenSSL
located in third_party/openssl.

-- Added OPENSSL argument to Makefile-Standalone that allows the developer to override the
above logic.

-- Updated Travis build script for OS X to use Makefile-Standalone, rather that calling
configure and make directly.

-- Added prebuild steps to Travis OS X builds to install homebrew packages for openssl,
automake and libtool.

-- Generally cleaned up the logic in Makefile-Standalone for handling OpenSSL options.

-- Cleaned up the Makefile-Standalone help text.

-- Fixed a small sed expression incompatibility in nrf5-app.mk that preventing building nRF5
SDK applications on OS X.