Fix a delayed resource deallocation
One of the fault injection tests was failing on weave-ping in a case
where a WRMP echo request was responded to but not acknowledged.  The
proper behavior for the echo client is to abort the exchange (which
implies not worrying about the pending retransmissions) when the valid
echo response was received and there is no pending action on a WRMP
ack.