Fix a test edge case in working with non-persistent counters
When using the non-persistent counters, the counters, by default,
start at 0; that behavior is exploited in a number of unit tests to
verify the event ID progression.  However, this leads to a delay in
event delivery in a case where only a single event for a particular
importance level has been logged; a number of tests that use fault
injection depend on a single event being delivered immediately.  Prior
to the refactoring of the initialization of the logging subsystem this
issue was addressed by passing an array of non-persistent counters
initialized to the current system time.  This patch restores that
functionality though at a different level: we add a method to the
LoggingManagement class to re-initialize the non-persistent counters
to the current system time.  This method is intended for use only in
test code.