Changed the initial value of the non-persistent logging counters.
The logging system treats the counter value `0` as special. We change
the starting value of non-persistent counters to `1`, and modify the
unit tests to reflect this assumption in the code.

This change also enabled us to remove the entire call chain to enable
an option to initialize the event IDs based on the timestamp.