Support for installable Python package (wheel) for OpenWeave
-- Added support for building an installable Python wheel (.whl) file that can be used to
install the OpenWeave Python code using pip on any Linux or OS X machine.  The OpenWeave
installable package includes the “openweave” Python module (containing WeaveDeviceManager
and other classes), the associated _WeaveDeviceMgr.so shared library, and the Weave Device
Manager Shell script (weave-device-mgr).  Once installed, the code and tools can be used
without reference to the openweave-core source tree, and without concern as to how the code
locates the shared library.

Building the OpenWeave Python installer requires the Python setuptools and wheel packages to
be installed on the build machine.  The --(enable|disable)-installable-python-package
configure option can be used to control whether the installable package is built.  By
default, the installable package will be built if configure detects that the prerequisite
build packages are available.