Ensure that N2J_NetworkInfo uses Make function with Channel and PAN ID
Currently this JNI conversion is using the NetworkInfo.Make() overload which does not take in a channel and PAN ID. Unfortunately, JNI doesn't complain if you pass too many arguments with the invocation, so the channel and PAN ID were being ignored. Additionally, although the ThreadChannel is only a uint_8, the method is declared in Java as an integer, so changed the cast to a jint instead.