Fix for Issue #283: Invalid Weave BLE scan response from nRF5 Device Layer
-- Corrected a mistake in the encoding of the Weave BLE device identification block that
gets sent as BLE service advertisement data by the nRF52840.  The previous version of the
code omitted two fields, which caused considerable heartburn for the Nest mobile
application.

-- Corrected another minor mistake in the encoding of the device identification block on the
ESP32 which resulted in the wrong data block type being sent.  This bug didn’t seem to
bother the mobile application.

-- Created a common data structure for encoding and decoding a Weave BLE device
identification block and refactored the nRF5 and ESP32 code bases to use this structure.