Corrected memory leak from extra memory allocation.
We were allocating twice the memory needed to buffer the QR code string. The second allocation (strdup) was overwriting the assignment of the pointer to the first. Consequently, the first allocation was never free'd. In the end, we don't need to allocate anything ourself and can defer to Apple's Foundation to do that for us.