Populate 802.15.4 MAC address on Thread-enabled platforms
On platforms that enable Thread, populate 802.15.4 extended
address. This address may be a meaningless set of numbers -- there is
no guarantee that the address populated will be the same as the
address used in joining (or re-attaching) to any particular Thread
network or that it will correspond to the factory provisioned EUI64.