Name an empty union to build with clang 9
This empty anonymous union builds correctly with clang 8, but clang 9
would emit a warning 'declaration does not declare anything', and would
prevent weave from building because we have '-Werror' on.

To keep the interface consistent, the union is temporarily named rather
than removed. Please be sure to remove the name when adding members.