Remove references to java.xml.bind.DataTypeConverter
java.xml.bind.DataTypeConverter has been deprecated in Java 9 and
removed in Java 11.  Our usage was minimal and limited to test code,
so we replace two methods from the DataTypeConverter --
`printHexString` and `parseHexString` -- with naiive implementations.
Base64 decoding has been replaced with the standard (as of Java 8)
java.util.Base64.