Update ESP32 Device Layer adaptation to use ESP-IDF 3.2
-- Removed reference to deprecated ESP AES function.

-- Changed logic around definition of LWIP_DNS_FOUND_CALLBACK_TYPE in DNSResolver code.

Previously, the logic surrounding the definition of LWIP_DNS_FOUND_CALLBACK_TYPE was written
with the presumption that the define existed in later versions of LwIP, and therefore only
needed to be declared in versions < 2.  However LWIP_DNS_FOUND_CALLBACK_TYPE has *never*
been a feature of stock LwIP.  Rather, it appears to be a Nest-ism intruduced in an internal
fork of LwIP. This change corrects the definition logic such that DNSResolver works with
the LwIP version included in ESP-IDF 3.2.

-- Ignore -Wdeprecated-declarations warnings when calling esp_wifi_get/set_auto_connect().

Later versions of ESP-IDF mark esp_wifi_get_auto_connect() and esp_wifi_set_auto_connect()
as depricated, leading to compilation errors.  Unfortunately, no obvious alternative exists,
and the comments and commit message provide no guidence.  A inquiry has been made to the
Expressif team.  Pending their answer, the decision has been made to continue using the
existing APIs.