Fix for Issue #346 : Provisioning hash logging feature
-- Added a method to the Device Layer ConfigurationManager interface to compute a
provisioning hash value for the device's provisioning data.  The output of the method is a
SHA256 hash value that confirms to the hash algorithm described in the Nest Weave: Factory
Provisioning Specification.

-- Added code to GenericConfigurationManagerImpl<> to implement the new
ComputeProvisioningHash() method.

-- Added a Device Layer configuration option (WEAVE_DEVICE_CONFIG_LOG_PROVISIONING_HASH)
that causes the device’s provisioning hash to be logged at boot time.

-- Fixed a rounding bug in the BASE64_ENCODED_LEN macro that overstated the expected length
of base-64 encoded data.

-- Added ESP32 Kconfig option for enabling provisioning hash logging.