Yet more fixes to simnet
-- Fixed support for tap interfaces within virtual nodes.  Unlike a veth-pair, the two ends
of a linux tap device cannot be in different namespaces.  To make it possible for a tap
interface in a node namespace to be connected to bridge in a “network” namespace, simnet now
constructs a “trunk line” veth-pair that spans the two namespaces, and a “tap bridge” in the
node namespace.  The tap interface is connected to the tap bridge, along with one of the
trunk interfaces.  The other trunk interface is connected to the bridge in the “network”
namespace.  This allows packets to flow between the process holding the tap fd and the
virtual network.

-- Eliminated the use of special short interface names when creating tap interfaces.  These
interfaces now use the names “wifi”, “thread”, etc., just like normal interfaces.

-- Disabled automatic assignment of IPv6 link-local addresses on interfaces that don’t need
them (e.g. the bridge interfaces in “network” namespaces, and the ends of a veth-pair that
are plugged into bridges).

-- Simplified the two-device simnet layout.