Fix a logic bug and perform minor cleanup
Based on the feedback from @lcniles, this patch cleans up a logic bug
in WeaveTimeClient.  WeaveTimeClient has a notion of "normal" state, I
ended up adopting "operational" state as the term.  The term denotes
all possible states that are not a part of initialization or shutdown
of the module.  This PR defines the canonical helper function,
replaces the faulty logic in `HandleTimeChangeNotification`, and
cleans up the correct but confusing logic in `Abort`.