Binding enhancements for connection-oriented WDM
-- Various changes to Binding class in preparation for full support of WDM over connection-
oriented transports.

-- Revised the binding reset logic to preserve the configuration state of the binding while
in the Failed state.  This allows code in the application’s BindingFailed handler to see the
original binding configuration values, such as the peer’s node id.

-- Modified Binding::IsAuthenticMessageFromPeer() to assert that an incoming message must be
received over the same transport as the binding.  In the case of a connection, the message
must have been received over the specific connection associated with the binding.

-- Simplified the interface to the Binding::Configuration::ConfigureFromMessage() method and
updated the logic to properly support messages received over a Weave connection.

-- Added API for querying transport type.