Relocate the test-driver script to clarify its ownership
weave-parallel-test-engine was originally based on the autotools'
test-driver.  Weave's version of the script added better signal
handling as well as a few tweaks to better support parallel test
execution.  Regrettably, these tweaks are quite specific to Weave, and
as a result awkward to upstream.  This commit relocates the script
third_party/autoconf/repo, a location that would logically closely
correspond to the parent, `test-driver`, script.