Python 3 migration -- Weave Device Manager
This commit adds a wrapper between C strings and Python strings.  We
wrap every argument specified to be of type `c_char_p` in a call to
`_StringToCString` and wrap every result of type `c_char_p` in a call
to `_CStringToString`.  Additional check is made against fixed-size,
NUL terminated `char` arrays (i.e. `c_char * N`).

binascii.hexlify has changed signature -- in Python 3 it returns an
array of ascii bytes, which needs to be decoded into a string.

Additionally, we update the singleton decorator to be python 3
compatible -- the previous version yielded an error
`WeaveDeviceManager is not callable`