Make the software update timer oblivious to service connectivity events
The existing behavior reset the software timers based on connectivity
to the service: when connectivity to the service was lost, the exiting
timer was cancelled, and when connectivity to the service was
restored, the idle timer was restored.  The intended effect was to
avoid software update checks when they were doomed to fail (service
unreachable) while preventing "thundering herd" behavior on service
outages.  However, the actual most common behavior was to reset the
long-running timer on any flapping in Thread network connectivity.

The solution is to just use the software update timer policy,
oblivious to the service connectivity.  The new code fails eagerly in
`PrepareBinding` when there is no service connectivity, which
generates no useless network activity.  The platform events are only
used to track whether we have a chance to succeed in contacting the
service.  Existing behavior already addresses the group behavior in
the event of mass service recovery -- the timer policy implements a
long, randomized background check, followed by randomized, exponential
backoff retry that ultimately reverts to the background rate.  In the
event of a synchronized start of a population of devices -- e.g. power
outage for line powered devices -- existing behavior also ensures
protection againts "thundering herds".